requirements.txt
torch
torchvision
onnx
onnxruntime
scikit-learn
pandas
numpy
matplotlib
fastapi
uvicorn
pyyaml 
seaborn 
